Added disassembler opcode until and including JUMPI

+++ disassembler/src/errors.rs
@@ -1,9 +1,9 @@
 use thiserror::Error;
 
-#[derive(Error)]
+#[derive(Error, Debug, PartialEq)]
 pub enum Error {
     #[error("Found unknown opcode: {0}")]
-    UnknownOpcodeError(u16),
+    UnknownOpcodeError(u32),
     #[error("Failed to read ROM file")]
     FileReadError,
     #[error("Failed to write to assembly file")]
